requirements.txt
streamlit==1.37.1
requests>=2.31.0
pydantic>=2.7
python-dateutil>=2.8
